refactor: [Lv3] User, Schedule 간 1:N 관계 설정 및 로직 수정

+++ src/main/java/org/example/spartaschedule/repository/JdbcTemplateScheduleRepository.java
@@ -46,7 +46,7 @@ public class JdbcTemplateScheduleRepository implements ScheduleRepository {
         // 데이터 SET
         Map<String, Object> parameters =  new HashMap<>();
         parameters.put("todo", schedule.getTodo());
-        parameters.put("writer", schedule.getWriter()); // FIXME: 추후 작성자 ID로 변경
+        parameters.put("userId", schedule.getUserId());
         parameters.put("password", schedule.getPassword());
         parameters.put("createdAt", currentTime);
         parameters.put("updatedAt", currentTime);
@@ -56,59 +56,62 @@ public class JdbcTemplateScheduleRepository implements ScheduleRepository {
 
         // 새로운 응답 객체 생성 후, 반환
         return new ScheduleResponseDto(
-                scheduleId.longValue(), schedule.getTodo(), schedule.getWriter(), currentTime, currentTime);
+                scheduleId.longValue(), schedule.getTodo(), schedule.getUserId(), currentTime, currentTime);
     }
 
     /**
      * [Repo] 전체 일정 리스트를 조회하고 반환하는 메서드
-     * 수정일 내림차순 & 작성자명 오름차순 정렬
+     * 수정일 내림차순 정렬
      *
      * @return 일정 객체 리스트
      */
     @Override
-    public List<Schedule> findAllSchedules() {
-        String query = "SELECT * FROM schedule ORDER BY updated_at DESC, writer ASC";
+    public List<Schedule> findAllSchedules(Long userId) {
+        String query = "SELECT * FROM schedule WHERE user_id = ? ORDER BY updated_at DESC";
 
-        return jdbcTemplate.query(query, scheduleRowMapper());
+        return jdbcTemplate.query(query, scheduleRowMapper(), userId);
     }
 
     /**
      * [Repo] 단일 일정을 조회하고 반환하는 메서드
-     * @param id 일정 id
+     * @param scheduleId 일정 id
+     * @param userId 유저 id
      * @return 조회된 일정 객체가 담긴 Optional 객체
      */
     @Override
-    public Optional<Schedule> findScheduleById(Long id) {
-        String query = "SELECT * FROM schedule WHERE id = ?";
-        List<Schedule> result = jdbcTemplate.query(query, scheduleRowMapper(), id);
+    public Optional<Schedule> findScheduleById(Long userId, Long scheduleId) {
+        String query = "SELECT * FROM schedule WHERE id = ? AND user_id = ?";
+        List<Schedule> result = jdbcTemplate.query(query, scheduleRowMapper(), scheduleId, userId);
 
         return result.stream().findFirst();
     }
 
     /**
      * [Repo] 일정을 수정하는 메서드
-     * @param id 일정 id
+     * @param scheduleId 일정 id
+     * @param userId 유저 id
      * @param dto 사용자 요청으로 수정한 일정 요청 객체
      * @return 수정한 행(row)의 개수
      */
     @Override
-    public int updateSchedule(Long id, ScheduleRequestDto dto) {
-        String query = "UPDATE schedule SET todo = ?, writer = ?, updated_at = ? WHERE id = ?";
+    public int updateSchedule(Long userId, Long scheduleId, ScheduleRequestDto dto) {
+        String query = "UPDATE schedule SET todo = ?, updated_at = ? WHERE id = ? AND user_id = ?";
 
         return jdbcTemplate.update(query,
-                dto.getTodo(), dto.getWriter(), LocalDateTime.now(), id);
+                dto.getTodo(), LocalDateTime.now(), scheduleId, userId);
     }
 
     /**
      * [Repo] 일정을 삭제하는 메서드
-     * @param id 삭제하고자 하는 일정 id
+     * @param scheduleId 일정 id
+     * @param userId 유저 id
      * @return 삭제한 행(row)의 개수
      */
     @Override
-    public int deleteSchedule(Long id) {
-        String query = "DELETE FROM schedule WHERE id = ?";
+    public int deleteSchedule(Long userId, Long scheduleId) {
+        String query = "DELETE FROM schedule WHERE id = ? AND user_id = ?";
 
-        return jdbcTemplate.update(query, id);
+        return jdbcTemplate.update(query, scheduleId, userId);
     }
 
     /**
@@ -122,7 +125,7 @@ public class JdbcTemplateScheduleRepository implements ScheduleRepository {
                 return new Schedule(
                         rs.getLong("id"),
                         rs.getString("todo"),
-                        rs.getString("writer"),
+                        rs.getLong("user_id"),
                         rs.getString("password"),
                         rs.getTimestamp("created_at").toLocalDateTime(),
                         rs.getTimestamp("updated_at").toLocalDateTime()
